chore: moved components to own files
